refactor code and add validation
